Allow web clients to construct MQTT server address
When the service is running on the Pi, to web browsers the MQTT server will
look like it is running on the same host as the web server but with a different
port number. But from within the server's Docker container, the MQTT server
looks like it is running on a separate host (since each container has its own
address on the internal bridge network).  Therefore we can't just tell the
client to connect to the same MQTT host the server uses.

Add config options to deal with this situation by letting the client construct
the MQTT server address based on the web UI's URL.